Major refactor for Elements to behave more like a SDK. Migrate logic to call API to containers instead of Elements components
