+++ frontend/.eslintrc.js
@@ -15,6 +15,14 @@ module.exports = {
     'max-depth': ['warn', 7],
     'max-lines': ['warn', 750],
     'array-callback-return': 'warn',
-    complexity: ['warn', { max: 20 }]
+    complexity: ['warn', { max: 20 }],
+
+    'react/prop-types': ['warn'],
+    'react/no-unused-prop-types': ['warn']
+  },
+  settings: {
+    'import/resolver': {
+      'babel-module': {}
+    }
   }
 };
