+++ frontend/src/Elements/components/Common/Checkbox.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import styles from '../theme/Checkbox.scss';
+import styles from '../../theme/Checkbox.scss';
 
 class Checkbox extends React.PureComponent {
   render() {
