+++ frontend/src/Elements/components/Common/PoweredBy.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+const PoweredBy = props => <div>Powered By ClearGDPR</div>;
+
+PoweredBy.propTypes = {
+  size: PropTypes.string
+};
+
+export default PoweredBy;
+
