--- frontend/src/containers/App.test.js
@@ -1,5 +1,0 @@
-// import { shallow } from 'enzyme';
-
-describe('(Component) App', () => {
-  it('exists', () => {});
-});
