--- frontend/src/containers/Profile.test.js
@@ -1,5 +1,0 @@
-// import { shallow } from 'enzyme';
-
-describe('(Component) Profile', () => {
-  it('exists', () => {});
-});
