--- frontend/src/containers/SignUp.test.js
@@ -1,5 +1,0 @@
-// import { shallow } from 'enzyme';
-
-describe('(Component) SignUp', () => {
-  it('exists', () => {});
-});
