+++ frontend/src/elements/index.js
@@ -1,10 +1,3 @@
-// import CGProvider from './components/Provider';
-// import injectCG from './components/inject';
-import Elements from './components/Elements';
-// import Element from './components/Element';
+import Elements from './Elements';
 
-export {
-  // CGProvider, injectCG,
-  Elements
-  // exampleElement
-};
+export default Elements;
