--- frontend/src/theme/demo.scss
@@ -1,8 +1,0 @@
-.toc {
-  background: linear-gradient(to bottom right, #22d0b2, #a2f4e3);
-  box-shadow: 0px 3px 10px rgba(0, 0, 0, 0.2);
-
-  a {
-    color: black;
-  }
-}
